CLOUD-5212: More logs
VMS-12734: new device added
VMS-18765: Fix signature crash on a garbage data
VMS-19217: Advanced PTZ mapping for MS-C2961-REPB.
VMS-19233: Wave server crashing once per several minutes
Branch: vms_4.1

+++ vms/libs/common/src/export/sign_helper.cpp
@@ -146,12 +146,12 @@ void QnSignHelper::doUpdateDigest(AVCodecID codecId, const quint8* extradata, in
         {
             while (curNal < dataEnd - reqUnitSize)
             {
-                int curSize = 0;
+                uint32_t curSize = 0;
                 for (int i = 0; i < reqUnitSize; ++i)
                     curSize = (curSize << 8) + curNal[i];
                 curNal += reqUnitSize;
-                curSize = qMin(curSize, (int)(dataEnd - curNal));
-                ctx.addData((const char*)curNal, curSize);
+                curSize = qMin(curSize, (uint32_t) (dataEnd - curNal));
+                ctx.addData((const char*) curNal, (int) curSize);
 
                 curNal += curSize;
             }
